requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
stripe==14.2.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==202601
uritemplate==4.2.0
urllib3==2.6.3
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
